Remove Grid element compile warnings on MacOS X gcc version 3.1.  MacOS X thinks timeval members tv_sec and tv_usec are ints (int_32_t), BSD & Linux think they are longs, MacOS complains when using timeval members in printfs, fix by casting to long.

+++ elements/grid/gridprobereplyreceiver.cc
@@ -90,14 +90,14 @@ GridProbeReplyReceiver::simple_action(Packet *p)
 
   char buf[200];
   snprintf(buf, sizeof(buf),
-	   "dest=%s nonce=%u hop=%s hopcount=%u rtt=%lu.%06lu tx_time=%lu.%06lu act=%u d1=%u d2=%u",
+	   "dest=%s nonce=%u hop=%s hopcount=%u rtt=%ld.%06ld tx_time=%ld.%06ld act=%u d1=%u d2=%u",
 	   IPAddress(rr->probe_dest).s().cc(),
 	   (unsigned int) ntohl(rr->nonce),
 	   IPAddress(gh->ip).s().cc(),
 	   (unsigned int) rr->reply_hop,
-	   rtt.tv_sec, rtt.tv_usec,
-	   tx_time.tv_sec,
-	   tx_time.tv_usec,
+	   (long) rtt.tv_sec, (long) rtt.tv_usec,
+	   (long) tx_time.tv_sec,
+	   (long) tx_time.tv_usec,
 	   (unsigned int) ntohl(rr->route_action),
 	   (unsigned int) ntohl(rr->data1),
 	   (unsigned int) ntohl(rr->data2));
